src/raw_to_embedding/requirements.txt
pymupdf
requests
beautifulsoup4
pydantic
python-dotenv
openai
numpy
faiss-cpu
sentence-transformers
